Adapters
In order for this circuit breaker to be used in distributed systems,
there needs to be a persistence store that can be read from any node.

This PR checks in the initial commit that makes the breaker usable via
Redis.  In order to use it, you'll need 3 arguments:

- adapter name - ie :redis
- adapter client - ie an instance of Redis
- adapter namespace - ie some unique value (maybe your method name?)
  that can be referenced

There's still more to do here with validations but at this point it's
working.  Also updated the example to showcase it.